+++ CerbySwapV1_Math.sol
@@ -3,16 +3,10 @@
 pragma solidity ^0.8.11;
 
 abstract contract CerbySwapV1_Math {
-
-
-    function sqrt(uint y) 
-        internal 
-        pure 
-        returns (uint z) 
-    {
+    function sqrt(uint256 y) internal pure returns (uint256 z) {
         if (y > 3) {
             z = y;
-            uint x = y / 2 + 1;
+            uint256 x = y / 2 + 1;
             while (x < z) {
                 z = x;
                 x = (y / x + x) / 2;
