+++ interfaces/IERC20.sol
@@ -2,13 +2,6 @@
 
 pragma solidity ^0.8.11;
 
-
 interface IERC20 {
-
-    function balanceOf(
-        address account
-    )
-        external
-        view
-        returns (uint);
+    function balanceOf(address account) external view returns (uint256);
 }
