+++ openzeppelin/access/AccessControl.sol
@@ -79,11 +79,15 @@ abstract contract AccessControl is Context, IAccessControl, ERC165 {
         _;
     }
 
-
     /**
      * @dev Returns `true` if `account` has been granted `role`.
      */
-    function hasRole(bytes32 role, address account) public view override returns (bool) {
+    function hasRole(bytes32 role, address account)
+        public
+        view
+        override
+        returns (bool)
+    {
         return _roles[role].members[account];
     }
 
@@ -120,17 +124,20 @@ abstract contract AccessControl is Context, IAccessControl, ERC165 {
      *
      * - the caller must have ``role``'s admin role.
      */
-    function grantRole(bytes32 role, address account) public virtual override onlyRole(getRoleAdmin(role)) {
+    function grantRole(bytes32 role, address account)
+        public
+        virtual
+        override
+        onlyRole(getRoleAdmin(role))
+    {
         _grantRole(role, account);
     }
 
-    
     function grantRolesBulk(RoleAccess[] calldata roles)
         public
         onlyRole(ROLE_ADMIN)
     {
-        for(uint i = 0; i<roles.length; i++)
-        {
+        for (uint256 i = 0; i < roles.length; i++) {
             _setupRole(roles[i].role, roles[i].addr);
         }
     }
@@ -144,7 +151,12 @@ abstract contract AccessControl is Context, IAccessControl, ERC165 {
      *
      * - the caller must have ``role``'s admin role.
      */
-    function revokeRole(bytes32 role, address account) public virtual override onlyRole(getRoleAdmin(role)) {
+    function revokeRole(bytes32 role, address account)
+        public
+        virtual
+        override
+        onlyRole(getRoleAdmin(role))
+    {
         _revokeRole(role, account);
     }
 
@@ -162,10 +174,12 @@ abstract contract AccessControl is Context, IAccessControl, ERC165 {
      *
      * - the caller must be `account`.
      */
-    function renounceRole(bytes32 role, address account) public virtual override {
-        if (
-            account != _msgSender()
-        ) {
+    function renounceRole(bytes32 role, address account)
+        public
+        virtual
+        override
+    {
+        if (account != _msgSender()) {
             revert CanOnlyRenounceRoleForSelf();
         }
