+++ openzeppelin/token/ERC1155/extensions/ERC1155Supply.sol
@@ -14,19 +14,19 @@ import "../ERC1155.sol";
  * same id are not going to be minted.
  */
 abstract contract ERC1155Supply is ERC1155 {
-    mapping(uint => uint) internal _totalSupply;
+    mapping(uint256 => uint256) internal _totalSupply;
 
     /**
      * @dev Total amount of tokens in with a given id.
      */
-    function totalSupply(uint id) public view virtual returns (uint) {
+    function totalSupply(uint256 id) public view virtual returns (uint256) {
         return _totalSupply[id];
     }
 
     /**
      * @dev Indicates whether any token exist with a given id, or not.
      */
-    function exists(uint id) public view virtual returns (bool) {
+    function exists(uint256 id) public view virtual returns (bool) {
         return ERC1155Supply.totalSupply(id) > 0;
     }
 
@@ -37,20 +37,20 @@ abstract contract ERC1155Supply is ERC1155 {
         address operator,
         address from,
         address to,
-        uint[] memory ids,
-        uint[] memory amounts,
+        uint256[] memory ids,
+        uint256[] memory amounts,
         bytes memory data
     ) internal virtual override {
         super._beforeTokenTransfer(operator, from, to, ids, amounts, data);
 
         if (from == address(0)) {
-            for (uint i = 0; i < ids.length; ++i) {
+            for (uint256 i = 0; i < ids.length; ++i) {
                 _totalSupply[ids[i]] += amounts[i];
             }
         }
 
         if (to == address(0)) {
-            for (uint i = 0; i < ids.length; ++i) {
+            for (uint256 i = 0; i < ids.length; ++i) {
                 _totalSupply[ids[i]] -= amounts[i];
             }
         }
